server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
